+++ tokio-threadpool/src/lib.rs
@@ -20,15 +20,14 @@ pub mod park;
 mod builder;
 mod callback;
 mod config;
-mod inner;
 #[cfg(feature = "unstable-futures")]
 mod futures2_wake;
 mod notifier;
+mod pool;
 mod sender;
 mod shutdown;
 mod shutdown_task;
 mod sleep_stack;
-mod pool_state;
 mod task;
 mod thread_pool;
 mod worker;
